pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackerdogs-tools"
version = "0.1.0"
description = "LangChain, CrewAI, and MCP Server Tools for hackerdogs.ai Platform"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "HackerDogs", email = "info@hackerdogs.ai"}
]
keywords = [
    "langchain",
    "crewai",
    "mcp",
    "threat-intelligence",
    "cybersecurity",
    "tools",
    "agents",
    "ai"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking :: Monitoring",
]

dependencies = [
    "aiofiles>=25.1.0",
    "annotated-types>=0.7.0",
    "anyio>=4.12.0",
    "backoff>=2.2.1",
    "beautifulsoup4>=4.14.3",
    "certifi>=2025.11.12",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.4",
    "click>=8.3.1",
    "colorlog>=6.10.1",
    "cryptography>=46.0.3",
    "dataclasses-json>=0.6.7",
    "easyocr>=1.7.2",
    "emoji>=2.15.0",
    "et-xmlfile>=2.0.0",
    "filelock>=3.20.0",
    "filetype>=1.2.0",
    "fsspec>=2025.12.0",
    "h11>=0.16.0",
    "hd-logging>=1.0.1",
    "html5lib>=1.1",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna>=3.11",
    "ImageIO>=2.37.2",
    "iniconfig>=2.3.0",
    "Jinja2>=3.1.6",
    "joblib>=1.5.2",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "docker>=6.0.0",  # For Docker client (optional, only if using Docker execution)
    "langdetect>=1.0.9",
    "langgraph>=1.0.4",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-prebuilt>=1.0.5",
    "langgraph-sdk>=0.2.12",
    "langsmith>=0.4.53",
    "lazy-loader>=0.4",
    "lxml>=6.0.2",
    "MarkupSafe>=3.0.3",
    "marshmallow>=3.26.1",
    "mpmath>=1.3.0",
    "mypy-extensions>=1.1.0",
    "narwhals>=2.13.0",
    "networkx>=3.6",
    "ninja>=1.13.0",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "olefile>=0.47",
    "opencv-python>=4.12.0.88",
    "opencv-python-headless>=4.12.0.88",
    "openpyxl>=3.1.5",
    "orjson>=3.11.4",
    "ormsgpack>=1.12.0",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "patsy>=1.0.2",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "plotly-express>=0.4.1",
    "pluggy>=1.6.0",
    "psutil>=7.1.3",
    "pyclipper>=1.4.0",
    "pycparser>=2.23",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "Pygments>=2.19.2",
    "pypdf>=6.4.0",
    "PyPDF2>=3.0.1",
    "pytesseract>=0.3.13",
    "pytest>=9.0.1",
    "python-bidi>=0.6.7",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-iso639>=2025.11.16",
    "python-magic>=0.4.27",
    "python-oxmsg>=0.0.2",
    "python-ulid>=3.1.0",
    "pytz>=2025.2",
    "PyYAML>=6.0.3",
    "RapidFuzz>=3.14.3",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "scikit-image>=0.25.2",
    "scipy>=1.16.3",
    "setuptools>=80.9.0",
    "shapely>=2.1.2",
    "six>=1.17.0",
    "soupsieve>=2.8",
    "statsmodels>=0.14.5",
    "sympy>=1.14.0",
    "tenacity>=9.1.2",
    "tifffile>=2025.10.16",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.2",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.2",
    "unstructured>=0.18.21",
    "unstructured-client>=0.42.4",
    "urllib3>=2.5.0",
    "uuid-utils>=0.12.0",
    "webencodings>=0.5.1",
    "wrapt>=2.0.1",
    "xxhash>=3.6.0",
    "zstandard>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
ti = [
    "OTXv2",
    "pycti",
    "pymisp",
]
osint = [
    # Identity tools (already Python packages)
    "sherlock-project",
    "maigret",
    "ghunt",
    "holehe",
    # Infrastructure tools (Python alternatives)
    "sublist3r",
    "dnsrecon",
    "dnspython",
    "python-nmap",
    "theHarvester",
    # Content tools
    "scrapy",
    "waybackpy",
    "internetarchive",
    "requests[socks]",
    "pysocks",
    "stem",
    # Metadata tools
    "exifread",
    "piexif",
    "yara-python",
    # Threat intelligence (already in ti)
    "OTXv2",
    "pymisp",
]
osint-binaries = [
    # Optional: Binary tools for maximum performance
    # These require system-level installation (brew/apt)
    # amass, nuclei, subfinder, masscan, zmap, waybackurls, exiftool, yara
]
all = [
    "hackerdogs-tools[dev,ti,osint]",
]

[project.urls]
Homepage = "https://github.com/hackerdogs/hackerdogs-tools"
Documentation = "https://github.com/hackerdogs/hackerdogs-tools#readme"
Repository = "https://github.com/hackerdogs/hackerdogs-tools"
Issues = "https://github.com/hackerdogs/hackerdogs-tools/issues"

[tool.setuptools]
packages = ["hackerdogs_tools", "hackerdogs_tools.ti", "hackerdogs_tools.prodx"]

[tool.setuptools.package-data]
hackerdogs_tools = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests", "hackerdogs_tools"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

